// src/api/api.ts
export interface Duration {
    'hours': number;
    'minutes': number;
}
export interface ProjectAllocationInner {
    'label': ProjectAllocationInnerLabelEnum;
    'percentage': number;
}

export const ProjectAllocationInnerLabelEnum = {
    It: 'IT',
    Hr: 'HR',
    Sales: 'Sales',
    Finance: 'Finance',
    Rd: 'R&D'
} as const;

export type ProjectAllocationInnerLabelEnum = typeof ProjectAllocationInnerLabelEnum[keyof typeof ProjectAllocationInnerLabelEnum];

export interface TimeChange {
    'id': number;
    'start': string;
    'end': string;
    'workTime': Duration;
    'pauseTime': Duration;
    'projectAllocation': Array<ProjectAllocationInner>;
    'workplaceAllocation': Array<WorkplaceAllocationInner>;
}
export interface WorkplaceAllocationInner {
    'label': WorkplaceAllocationInnerLabelEnum;
    'percentage': number;
}

export const WorkplaceAllocationInnerLabelEnum = {
    MobileWork: 'Mobile Work',
    InMotion: 'in-motion',
    Customer: 'Customer'
} as const;

export type WorkplaceAllocationInnerLabelEnum = typeof WorkplaceAllocationInnerLabelEnum[keyof typeof WorkplaceAllocationInnerLabelEnum];